pyproject.toml
[tool.poetry]
name = "ankh"
version = "0.2.0"
description = "A block cipher using a HMAC feistel network"
authors = ["Osiriz <0siriz@proton.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pycryptodome = "^3.20.0"
numpy = "^1.26.4"


[tool.poetry.group.test.dependencies]
pytest = "^8.1.1"
pytest-xdist = "^3.6.1"


[tool.poetry.group.dev.dependencies]
ipython = "^8.25.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
